+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Version 0.1.1
-- Feature
+- Features
   + Version tracker
 - Enhancements
   + User count on Group list
   + Time on Message (primitive)
-- Bug Fixes:
+- Bug Fixes
   + Routing: Allow refresh
   + Routing: Allow deep linking
+
+# Version 0.1.2
+- Enhancements
+  + Message design updates
+  + Time on Message (HH:MM:SS)
+- Regression
+  + Possible: Message input lag